Update dependencies and fix clippy warnings.

+++ src/helpers.rs
@@ -13,7 +13,7 @@ where
     let prev = index as usize % arr.len();
     let next = (prev + 1) % arr.len();
     // todo: make this look cleaner
-    ((arr[next].clone().into() - arr[prev].clone().into()) * (index as f32 - prev as f32)
+    ((arr[next].clone().into() - arr[prev].clone().into()) * (index - prev as f32)
         + arr[prev].clone().into())
     .into()
 }
@@ -28,7 +28,7 @@ pub fn semitone_from_frequency(frequency: f32) -> u8 {
 }
 
 pub fn frequency_from_semitone(semitone: f32) -> f32 {
-    let diff = semitone as f32 - 57.0;
+    let diff = semitone - 57.0;
     let ratio = 2.0_f32.powf(1.0 / 12.0);
     MID_A_FREQUENCY * ratio.powf(diff)
 }
